requirements.txt
# 数据源
tushare==1.4.4

# 数据处理
pandas==2.1.3
numpy==1.26.2

# Web框架
flask==3.0.0
flask-cors==4.0.0

# 日志
loguru==0.7.2

# 配置
pyyaml==6.0.1

# 其他工具
requests==2.31.0
